fix #61
