+++ src/env.rs
@@ -7,6 +7,8 @@ use std::env;
 use std::process::Command;
 
 use crate::env::dep_source_replace::filter_cargo_tree;
+use crate::time::now_data_time;
+use crate::Format;
 use std::collections::HashMap;
 
 #[derive(Default, Debug)]
@@ -249,20 +251,6 @@ const BUILD_TIME_2822: ShadowConst = "BUILD_TIME_2822";
 const BUILD_TIME_3339: ShadowConst = "BUILD_TIME_3339";
 const BUILD_RUST_CHANNEL: ShadowConst = "BUILD_RUST_CHANNEL";
 
-fn now_data_time() -> DateTime<Local> {
-    match env::var_os("SOURCE_DATE_EPOCH") {
-        None => Local::now(),
-        Some(timestamp) => {
-            let epoch = timestamp
-                .into_string()
-                .expect("Input SOURCE_DATE_EPOCH could not be parsed")
-                .parse::<i64>()
-                .expect("Input SOURCE_DATE_EPOCH could not be cast to a number");
-            Local.timestamp(epoch, 0)
-        }
-    }
-}
-
 pub fn build_time(project: &mut Project) {
     // Enable reproducible builds: https://reproducible-builds.org/docs/source-date-epoch/
     let time = now_data_time();
@@ -270,7 +258,7 @@ pub fn build_time(project: &mut Project) {
         BUILD_TIME,
         ConstVal {
             desc: "display project build time".to_string(),
-            v: time.format("%Y-%m-%d %H:%M:%S").to_string(),
+            v: time.human_format(),
             t: ConstType::Str,
         },
     );
@@ -379,13 +367,5 @@ mod tests {
         let input = r#"shadow-rs v0.5.23 (FD:\a\shadow-rs\shadow-rs)"#;
         let ret = filter_dep_source(input);
         assert_eq!(input, ret)
-    }
-
-    #[test]
-    fn test_now_data_time() {
-        std::env::set_var("SOURCE_DATE_EPOCH", "1628080443");
-        let time = now_data_time();
-        let now = Local::now();
-        assert!(time < now);
     }
 }
